Get all log events.
DescribeLogStreams will only bring back a maximum of 50 streams. For a
large number of files, there may be more than this, especially for the
file format lambda which produces a lot of logs.

We need to recursively loop around until the nextToken is blank so we're
sure we've got them all.

I've also added `take(10)` to the report so it doesn't print more than
10 records to the web page.